MOD: Victoria?? Eso espero
En fin, que la escala era un entero xDDD

+++ Practica1-OHNO/AndroidEngine/src/main/java/es/ucm/gdv/blas/oses/carreau/androidengine/Android/AndroidGraphics.java
@@ -49,12 +49,12 @@ public class AndroidGraphics extends AbstractGraphics {
 
     ////////////////////METODOS DE TRANSFORMACION//////////////////////////////
     @Override
-    public void translate(int x, int y) {
+    public void translate(float x, float y) {
         this.canvas.translate(x, y);
     }
 
     @Override
-    public void scale(int w, int h) {
+    public void scale(float w, float h) {
         this.canvas.scale(w, h);
     }
